separated public and private ledger stuff
