+++ rippled/transaction.js
@@ -1,110 +1,14 @@
 /**
  * http://usejsdoc.org/
  */
-const walletStore = require('../wallet/walletStore')
-const Decimal = require('decimal.js')
 const Amount = require('./amount').Amount
 const source = require('./source')
 const destination = require('./destination')
-const specification = require('./specification')
 const hpc = require('../utils/hastyPuddingCipherUtil')
 const RippleAPI = require('ripple-lib').RippleAPI
 const api = new RippleAPI({
 	server: 'wss://s.altnet.rippletest.net:51233'
 })
-
-
-//TODO consider re-authentication when moving lots of $$$ https://www.owasp.org/index.php/Transaction_Authorization_Cheat_Sheet	
-	
-function tipUser(sourceUserID, destinationUserID, amount) {
-	return new Promise(function(resolve, reject) {
-		walletStore.getUserBalance(sourceUserID).then((balance) => {
-			let amountAsDecimal = new Decimal(amount)
-			let balanceAsDecimal = new Decimal(balance)
-			
-			// check that the tipper has enough funds to actually tip
-			if(balanceAsDecimal.lessThan(amountAsDecimal)){
-				reject('Insufficient funds to tip ' + amount + 'XRP')
-			}
-			else{
-				//TODO need error handling here and to roll back first transaction should second fail
-				//mayb there is a way to do transactions in batches
-				let tipperNewBalance = balanceAsDecimal.sub(amountAsDecimal)
-				walletStore.updateUserBalance(sourceUserID, tipperNewBalance.toString()).then((result) => {
-					if(result){
-						walletStore.getUserBalance(destinationUserID).then((balance) => {
-							let tippeeBalance = new Decimal(balance)
-							let tippeeNewBalance = tippeeBalance.add(amountAsDecimal)
-							console.log("tippeeNewBalance", tippeeNewBalance.toString())
-							walletStore.updateUserBalance(destinationUserID, tippeeNewBalance.toString()).then((result) => {
-								if(result){
-									resolve('Tipped ' + amount + 'XRP')
-								}
-								else{
-									//TODO need to roll everything back
-								}
-							})
-							
-						})
-					}
-					else{
-						console.log('error: ' + result)
-					}
-				})
-				/*
-				let amountAsDrops = amountAsDecimal.mul(100000) //convert XRP to drops
-				let tipAmount = new Amount(amountAsDrops.toString())
-				let s = source.buildSource(daoAddress, tipAmount, Number(hpc.map(sourceUserID).toString(10).substring(0, 10)))
-				let d = destination.buildDestination(daoAddress, tipAmount, Number(hpc.map(destinationUserID).toString(10).substring(0, 10)))
-				let payment = specification.buildSpecification(s, d)*/
-				/*
-				 * We want payment objects to look like this:
-				 * 
-				 * payment = { 
-					"source": { 
-						"address": "rwYQjHp9HZiKKpZB4i4fvc8eQvAtA7vdY6", 
-						"tag": 1176135248, 
-						"maxAmount": { 
-							"currency": "drops", 
-							"value": "1000000" 
-						} 
-					}, 
-					"destination": { 
-						"address": "rwYQjHp9HZiKKpZB4i4fvc8eQvAtA7vdY6", 
-						"amount": { 
-							"currency": "drops", 
-							"value": "1000000" 
-						}, 
-						"tag": 1530675606
-					} 
-				}
-				
-				payment = {
-						  "source": {
-							    "address": "r9cZA1mLK5R5Am25ArfXFmqgNwjZgnfk59",
-							    "maxAmount": {
-							      "value": "0.01",
-							      "currency": "USD",
-							      "counterparty": "rMH4UxPrbuMa1spCBR98hLLyNJp4d8p4tM"
-							    }
-							  },
-							  "destination": {
-							    "address": "rpZc4mVfWUif9CRoHRKKcmhu1nx2xktxBo",
-							    "amount": {
-							      "value": "0.01",
-							      "currency": "USD",
-							      "counterparty": "rMH4UxPrbuMa1spCBR98hLLyNJp4d8p4tM"
-							    }
-							  }
-							}
-				*/
-				//preparePaymentTransaction(daoAddress, payment)
-			}
-		})
-	})
-	
-	
-}
 
 // BEGIN TRANSACTION FLOW METHODS
 //TODO these are nested calls and problems can occur if the api connection is severed too early. need to Un-nest
@@ -154,5 +58,7 @@ function verifyTransaction(transactionID){
 }
 // END TRANSACTION FLOW METHODS
 module.exports = {
-		tipUser
+		signTransaction,
+		submitTransaction,
+		verifyTransaction
 }
